[REFACTOR] data하위로 model, modelView 이전
