+++ app/src/main/java/com/refit/app/ui/composable/common/BottomBar.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import com.refit.app.R
-import com.refit.app.ui.composable.model.basic.BottomNavItem
+import com.refit.app.data.basic.model.BottomNavItem
 import com.refit.app.ui.theme.DarkBlack
 import com.refit.app.ui.theme.MainPurple
 import com.refit.app.ui.theme.Pretendard
