+++ app/src/main/java/com/refit/app/ui/composable/common/RefitTopBar.kt
@@ -1,8 +1,7 @@
 package com.refit.app.ui.composable.common
 
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
-import com.refit.app.ui.composable.model.basic.AppBarConfig
+import com.refit.app.data.basic.model.AppBarConfig
 import com.refit.app.ui.composable.common.topbar.variants.HomeSearchTopBar
 import com.refit.app.ui.composable.common.topbar.variants.SearchOnlyTopBar
 import com.refit.app.ui.composable.common.topbar.variants.HomeTitleTopBar
