+++ app/src/main/java/com/refit/app/ui/composable/common/topbar/variants/TopBarBackWithActions.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.refit.app.ui.composable.model.basic.AppBarConfig
+import com.refit.app.data.basic.model.AppBarConfig
 import com.refit.app.ui.composable.common.topbar.*
 
 @OptIn(ExperimentalMaterial3Api::class)
