+++ app/src/main/java/com/refit/app/ui/composable/common/topbar/variants/TopBarHomeTitle.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
-import com.refit.app.ui.composable.model.basic.AppBarConfig
+import com.refit.app.data.basic.model.AppBarConfig
 import com.refit.app.ui.composable.common.topbar.*
 
 @OptIn(ExperimentalMaterial3Api::class)
