+++ app/src/main/java/com/refit/app/ui/composable/common/topbar/variants/TopBarSearchOnly.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.refit.app.R
-import com.refit.app.ui.composable.model.basic.AppBarConfig
+import com.refit.app.data.basic.model.AppBarConfig
 import com.refit.app.ui.composable.common.topbar.TopBarTokens
 import com.refit.app.ui.theme.MainPurple
 import androidx.compose.ui.text.input.ImeAction
