+++ app/src/main/java/com/refit/app/ui/composable/health/DailyList.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import com.refit.app.model.health.DailyRow
+import com.refit.app.data.health.model.DailyRow
 import java.time.format.DateTimeFormatter
 import kotlin.math.roundToInt
