+++ app/src/main/java/com/refit/app/ui/composable/home/MetricRow.kt
@@ -6,7 +6,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.refit.app.ui.composable.model.health.MetricItem
+import com.refit.app.data.health.model.MetricItem
 
 @Composable
 fun MetricRow(items: List<MetricItem>) {
